Changed buffer to uint16 as default mode is YUV422
Color values for one color pixel is presented in two consecutive uint16. Least
significant bits of every uint16 represents the luminance (brightness/grayscale)
of the pixel.

+++ include/dcmi_ov2640.h
@@ -38,7 +38,7 @@
 #include "stm32f4xx_i2c.h"
 #include "stm32f4xx_dcmi.h"
 
-extern unsigned int Targetbuffer[160*120];
+extern uint16_t Targetbuffer[160*120];
 /* Exported types ------------------------------------------------------------*/
 typedef struct
 {
